refactor: Factor out schema to amek it reusable for web
